Added Account Component (Resolved #67)
