+++ client/src/components/edit.component.js
@@ -9,7 +9,7 @@ const SERVER_URL = process.env.REACT_APP_SERVER_URL;
 const DATE_VAL = /^(19|20|21)\d\d[- /.](0[1-9]|1[012])[- /.](0[1-9]|[12][0-9]|3[01])$/;
 
 const Edit = ({ userData }) => {
-    const {email, id: userId, authenticated, isLoading} = userData;
+    const {id: userId, authenticated, isLoading} = userData;
     const {id} = useParams();
     const [title, setTitle] = useState('loading ...');
     const [date, setDate] = useState('2020-01-01');
